scClustBench

The data given by mat is clustered in method algorithm with various similarity metrics for benchmark.

The data mat must contain label information as column names to subsample the matrix by subset_p number of cells per cluster in each rep.

scClustBench(mat, nCs, method = "simlr", similarity = NULL,
  geneFilter = 0.8, rep = 5, subset_p = 0.8, cores = 1, seed = 1,
  ...)

Arguments

mat
a (m x n) data matrix of gene expression measurements of individual cells with rows representing genes and columns representing cells. column names of the mat must be cell types.

nCs
number of clusters to be estimated

method
Clustering method to be performed on the dataset between "simlr" from implemented version of SIMLR package or "kmeans" from amap package. It is set to "simlr" by default.

similarity
A vector of similarity metrics to be used for clustering.

geneFilter
A threshold to remove genes. The genes that are not expressed more than the threshold across all the cells in the dataset will be removed. Genes will not be removed if set to 0.

rep
A number of subsampling of the matrix

subset_p
Sampling percentage per cell types from mat

cores
Number of cores to be used for parallel processing

seed
seed for randomisation

...
An addtional paramaters for corresponding clustering method specified from method

Value

A list with length rep. Each item in the list contain a list object indexed by similarity and the true label ("truth"). The object indexed by similarity metric is a clustering result object.

Details

Benchmark impact of similarity metric on clustering.