Training scClassify model

Training scClassify model

train_scClassify(exprsMat_train, cellTypes_train, tree = c("HOPACH",
  "HC"), selectFeatures = c("limma", "DV", "DD", "chisq", "BI"),
  topN = 50, hopach_kmax = 5, pSig = 0.05, parallel = F,
  ncores = 1, verbose = T)

Arguments

exprsMat_train
A matrix of expression matrix of training dataset

cellTypes_train
A vector of cell types of training dataset

tree
A vector indicates the method to build hierarchical tree, set as "HOPACH" by default. This should be one of "HOPACH" and "HC" (using stats::hclust).

selectFeatures
A vector indicates the method to select features, set as "limma" by default. This should be one or more of "limma", "DV", "DD", "chisq", "BI".

topN
An integer indicates the top number of features that are selected

hopach_kmax
An integer between 1 and 9 specifying the maximum number of children at each node in the HOPACH tree.

pSig
A numeric indicates the cutoff of pvalue for features

parallel
A logical input indicates whether the algorihms will run in parallel

ncores
An integer indicates the number of cores that are used

verbose
A logical input indicates whether the intermediate steps will be printed

Value

list of results